+++ src/main/java/com/example/team1_be/domain/Schedule/ScheduleRepository.java
@@ -2,5 +2,5 @@ package com.example.team1_be.domain.Schedule;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScheduleRepository extends JpaRepository<Schedule, Integer> {
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
 }
