+++ src/main/java/com/example/team1_be/domain/Week/WeekRepository.java
@@ -2,6 +2,6 @@ package com.example.team1_be.domain.Week;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WeekRepository extends JpaRepository<Week, Integer> {
+public interface WeekRepository extends JpaRepository<Week, Long> {
 
 }
